requirements.txt
anthropic
openai
python-dotenv
httpx
pydantic
tenacity
rich
fastapi
uvicorn
duckduckgo-search
